Added filters.js Separated the existing script.js -> filtesr.js and script.js
